requirements.txt
contourpy==1.2.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.5.0
ftfy==6.2.0
imageio==2.34.1
jinja2==3.1.4
kiwisolver==1.4.5
lazy-loader==0.4
lightning-utilities==0.11.2
markupsafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.2
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
regex==2024.5.15
scikit-image==0.22.0
scipy==1.13.0
six==1.16.0
sympy==1.12
tifffile==2024.5.10
torch-fidelity==0.3.0
torchmetrics==1.4.0.post0
tqdm==4.66.4
typing-extensions==4.11.0
versioneer==0.29
wcwidth==0.2.13